Add a placeholder for display resolution
Previously, on mousedown on
slider the 16x16 would disappear
which shifted the slider up slightly.
This caused compatibility
issues in Safari (mouseup cannot
be off of line to register a full click
/ Chrome does allow you to lift off
of the element and it still
registers). Now a placeholder is temporarily added
and then removed so spacing is
preserved.